refactor(download): move files.py to download/client.py; Files -> Client
The facade moved to pridepy/download/client.py and the class renamed
Files -> Client. pridepy/files/files.py is now a back-compat shim
(Files = Client, plus Progress re-export). CLI imports Client (aliased
as Files internally to keep the diff minimal). Test imports keep working
via the shim. No behaviour change; 68 passed, 4 skipped.
